% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_geom.R
\name{calculate_geom}
\alias{calculate_geom}
\title{Calculate grid based on type and seed}
\usage{
calculate_geom(file, type, seed)
}
\arguments{
\item{file}{The file that has been loaded}

\item{type}{The type of grid to generate (square or hexagonal)}

\item{seed}{The seed to generate the grid}
}
\value{
The grid generated
}
\description{
Calculate grid based on type and seed
}
