% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_geom.R
\name{tidy_geom}
\alias{tidy_geom}
\title{Calculate grid based on type and seed}
\usage{
tidy_geom(geom)
}
\arguments{
\item{geom}{Geom generated by calculate_geom}
}
\value{
Tidy version of the geom
}
\description{
Calculate grid based on type and seed
}
\keyword{internal}
