% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\title{Function to get the corresponding grid for a map}
\usage{
get_grid(file, type = "hexagonal", seed = 2019)
}
\arguments{
\item{file}{The file to be loaded. Comes from st_read() function.}

\item{type}{The type of grid to generate (square or hexagonal)}

\item{seed}{The seed to generate the grid}
}
\value{
The grid generated
}
\description{
Function to get the corresponding grid for a map
}
